Isolate ThemeSwitcher component
